backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
boto3==1.35.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.4.0
scikit-image==0.24.0
httpx==0.27.2
python-dotenv==1.0.1
aiosmtplib==3.0.1
jinja2==3.1.4
